requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-allauth==0.41.0
django-import-export==2.0.2
django-suit==0.2.28
django-weasyprint==0.5.5
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.3.12
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==2.6.4
Pillow==7.0.0
pycparser==2.20
pycrypto==2.6.1
PyPDF2==1.26.0
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tablib==1.1.0
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
